Adding 'modified' alias
Adds the 'modified' alias to the collection. This function will list all files modified in the branch (as compared to the parent branch). If the branch to select is not specified, the current working branch is selected. If the parent/comparison branch is not specified, the parent (merge) branch of the selected branch is used. As with other aliases, pattern matching is used to determine the full branch name when a branch name is specified.